+++ utils.js
@@ -1,48 +1,18 @@
 const fs = require("fs");
-const moment = require("moment");
 const pdf = require("pdf-parse");
-
-/**
- * Returns a string representation of the week of a given date.
- * @param {Date} date - The date.
- * @return {string} - The string representation of the week (e.g., "Oct 03 - Oct 09, 2023").
-*/
-function getWeekString(date) {
-    const startOfWeek = moment(date).startOf('week').day(0); // Set week start on Sunday
-    const endOfWeek = moment(date).endOf('week').day(6); // Set week end on Saturday
-    return `${startOfWeek.format('MMM DD')} - ${endOfWeek.format('MMM DD, YYYY')}`;
-}
-
-/**
- * Helper function to parse date strings from the PDF into Date objects.
- * @param {string} dateStr - The date string from the PDF.
- * @return {Date} - A JavaScript Date object.
- */
-function parseDate(dateStr) {
-    const formattedDateStr = dateStr.replace(' at ', ' ').replace(/(AM|PM)/, ' $1');
-    return new Date(moment(formattedDateStr, 'MM/DD/YYYY hh:mm A').toDate());
-}
-
-/**
- * Helper function to format date strings from the PDF into a more readable format.
- * @param {string|object} dateStr - The parsed date object from the PDF, or a string ('Never')
- */
-function formatDate(dateStr) {
-    // console.log('dateStr', dateStr, typeof dateStr);
-    if (typeof dateStr === 'object') {
-        return dateStr.toString().replace(' GMT-0800 (Pacific Standard Time)', '');
-    }
-    return dateStr;
-}
-
+const {
+    getWeekString,
+    parseDate,
+    formatDate,
+} = require('./utils/date');
 
 /**
  * Writes the provided data to a file at the provided file path, and logs a
  * confirmation message to the console.
-*
-* @param {string} filePath - The path to the output file.
-* @param {string} data - The data to write to the file.
-*/
+ *
+ * @param {string} filePath - The path to the output file.
+ * @param {string} data - The data to write to the file.
+ */
 function writeFile(filePath, data) {
     try {
         fs.writeFileSync(filePath, data);
@@ -56,7 +26,7 @@ function writeFile(filePath, data) {
  * Parses the PDF file at the specified file path.
  * @param {string} filePath - The path to the PDF file.
  * @return {Promise<string>} - A promise that resolves to the text content of the PDF file.
-*/
+ */
 async function parsePdf(filePath) {
     try {
         const dataBuffer = fs.readFileSync(filePath);